Unerf aggressive
I had thought that the half a castle in gold could only be used by the end of turn 3. It turns out there are some openings where this gold is needed much sooner. The gold is now given immediately on phase 1. This is technically a buff on the Aggressive of Risky Business as there is now an option of rushing units, mines or a unit upgrade instead for Aggressive players.